UI fixes.
- searchField is cleared on begin editing, this doesn't trigger editingChanged.
- make service cells editable to enable swipe to delete.
- selectedService not updated when service table's selection changes.
- invisible actionStack inhibits access to newButton.

+++ Sources/View/MPEffectView.swift
@@ -189,7 +189,7 @@ class MPEffectView: UIView {
 
             self.blurEffectView.layer.cornerRadius = self.isRound ? min( self.bounds.width, self.bounds.height ) / 2: self.rounding
             self.blurEffectView.layer.borderColor = (self.borderColor ?? self.tintColor)?.cgColor
-            self.blurEffectView.alpha = self.isDimmed ? .long: .on
+            self.blurEffectView.alpha = self.isDimmed ? .short: .on
         }
     }
 }
